MS-939 Timer extracted for TimeHelper
