+++ feature/dashboard/src/test/java/com/simprints/feature/dashboard/tools/di/FakeCoreModule.kt
@@ -14,6 +14,7 @@ import com.simprints.core.PackageVersionName
 import com.simprints.core.SessionCoroutineScope
 import com.simprints.core.tools.json.JsonHelper
 import com.simprints.core.tools.time.TimeHelper
+import com.simprints.core.tools.time.Timer
 import com.simprints.core.tools.utils.EncodingUtils
 import com.simprints.core.tools.utils.StringTokenizer
 import com.simprints.testtools.unit.EncodingUtilsImplForTests
@@ -43,6 +44,10 @@ object FakeCoreModule {
     @Provides
     @Singleton
     fun provideTimeHelper(): TimeHelper = mockk(relaxed = true)
+
+    @Provides
+    @Singleton
+    fun provideTimer(): Timer = mockk()
 
     @Provides
     @Singleton
