+++ infra/core/src/main/java/com/simprints/core/tools/time/KronosTimeHelperImpl.kt
@@ -4,9 +4,6 @@ import android.text.format.DateUtils.FORMAT_SHOW_DATE
 import android.text.format.DateUtils.MINUTE_IN_MILLIS
 import android.text.format.DateUtils.getRelativeTimeSpanString
 import com.lyft.kronos.KronosClock
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import java.text.DateFormat
 import java.util.Calendar
 import java.util.Date
@@ -61,16 +58,5 @@ class KronosTimeHelperImpl @Inject constructor(
         set(Calendar.MILLISECOND, 0)
 
         timeInMillis
-    }
-
-    override fun observeTickOncePerMinute(): Flow<Unit> = flow {
-        while (true) {
-            emit(Unit)
-            delay(ONE_MINUTE_IN_MILLIS)
-        }
-    }
-
-    private companion object {
-        private const val ONE_MINUTE_IN_MILLIS = 60 * 1000L
     }
 }
