+++ infra/core/src/main/java/com/simprints/core/tools/time/Timer.kt
@@ -1,0 +1,9 @@
+package com.simprints.core.tools.time
+
+import androidx.annotation.Keep
+import kotlinx.coroutines.flow.Flow
+
+@Keep
+interface Timer {
+    fun observeTickOncePerMinute(): Flow<Unit>
+}
